+++ src/component/AppLayout/Welcome.jsx
@@ -19,11 +19,12 @@ export class Welcome extends AbstractComponent {
   }
 
   onOpenDemoProject = async ( e ) => {
-    const { loadProject } = this.props.action;
+    const { loadProject, saveSettings } = this.props.action;
     e.preventDefault();
     this.setState({ error: "" });
     try {
       const demoProject = await getDemoProjectDirectory();
+      demoProject && saveSettings({ projectDirectory: demoProject });
       demoProject && await loadProject( demoProject );
     } catch ( err ) {
       this.setState({ error: err.message });