+++ src/component/Modal/AbstractTestRunnerModal.jsx
@@ -5,8 +5,7 @@ import StateStorage from "service/StateStorage";
 export default class AbstractTestRunnerModal extends AbstractComponent {
 
   getBrowserOptions() {
-    const storage = new StateStorage( this.constructor.name );
-
+    const storage = new StateStorage( this.constructor.name, sessionStorage );
     // Browser options tab is not shown
     if ( this.refBrowserOptions.current === null ) {
 
@@ -14,12 +13,13 @@ export default class AbstractTestRunnerModal extends AbstractComponent {
         incognito: true,
         "puppeteer.connect": {
           ignoreHTTPSErrors: true,
-          slowMo: 30,
           browserWSEndpoint: null
         },
         "puppeteer.launch": {
+          args: [],
           devtools: false,
           headless: true,
+          slowMo: 30,
           ignoreHTTPSErrors: false
         }
       };