+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# Puppetry 3.2.2
+
+### Features
+- improvement: added "replace existing value" to target.type
+- improvement: added page.authenticate (issue #70)
+- improvement: added target.assertText (to compare text ignoring HTML/styling)
+
+### Bug fixes
+- fix: issues #71
+
+### Updates
+- improvement: update to Puppeteer 3.1.0
+
+### Bug fixes
+- improve error output in test report
+
 # Puppetry 3.2.1
 
 ### Bug fixes